debug tests

+++ pennylane/templates/state_preparations/partial_unary.py
@@ -780,7 +780,7 @@ def _pui_state_prep_core(coefficients, wires, indices, work_wires):
             k_start, k = data
             qp.BasisState(k_start, subspace_wires)
             b = k - k_start
-            qp.QROM(np.eye(b), subspace_wires, nonsubspace_wires[:b], work_wires[: n_subspace - 1])
+            qp.QROM(np.eye(b, dtype=np.int64), subspace_wires, nonsubspace_wires[:b], work_wires[: n_subspace - 1])
             qp.BasisState(k_start, subspace_wires)
             continue
         if _type == "Fanout":
